+++ CHANGELOG.md
@@ -11,9 +11,9 @@ All notable changes to this project will be documented in this file.
 
 ## 0.1.0 - 2015-04-08
 ### Added
-- Option to set up Docker Compose project name of Docker instances.
-- Option to set up Docker Compose configuration file path.
-- Option to set up Docker Compose command description.
+- Definition `name` for a name of instances.
+- Definition `file` for a path of Docker Compose configuration file.
+- Definition `description` for a command description.
 - Simple running Docker Compose without any options.
 
 [unreleased]: https://github.com/phpzone/docker/compare/0.1.1...HEAD